requirements.txt
playwright
prometheus_client
ultralytics
opencv-python-headless
numpy
pillow
paddlepaddle
paddleocr
tqdm
pytest
pydantic
langchain
langchain-community
langchain-core
langchain-openai
python-dotenv
